Restructured folder, add utils, fixed header
utils: add AutoScrollToTop
Header: Fix bug
