+++ src/App.jsx
@@ -1,30 +1,37 @@
-import './App.css'
+import './App.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import NavbarHeader from './components/NavbarHeader';
 import Footer from './components/Footer';
-import {BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Login from './page/Login';
-import Home from './page/Home'
-import Register from './page/Register';
-import Introduce from './page/Introduce';
-import Service from './page/service';
-import '@fortawesome/fontawesome-free/css/all.min.css';
+import { BrowserRouter as Router, Routes, Route, useLocation } from 'react-router-dom';
+import Login from './page/Login/Login'
+import Home from './page/home/Home';
+import Register from './page/Login/Register'
+import Introduce from './page/Introduce/Introduce';
 import Header from './components/Header';
+
+function AppContent() {
+  const location = useLocation();
+
+  return (
+    <>
+      <NavbarHeader />
+      {location.pathname !== '/home' && location.pathname !== '/' && <Header />}
+      <Routes>
+        <Route index element={<Home />} />
+        <Route path="/home" element={<Home />} />
+        <Route path="/register" element={<Register />} />
+        <Route path="/login" element={<Login />} />
+        <Route path="/introduce" element={<Introduce />} />
+      </Routes>
+      <Footer />
+    </>
+  );
+}
 
 export default function App() {
   return (
     <Router>
-      <NavbarHeader />
-      <Header />
-      <Routes>
-      <Route index element={<Home />} />
-      <Route path="/home" element={<Home />} />
-        <Route path="/register" element={<Register />} />
-        <Route path="/login" element={<Login />} />
-        <Route path="/introduce" element={<Introduce />} />
-        <Route path="/service" element={<Service />} />
-      </Routes>
-      <Footer />
+      <AppContent />
     </Router>
   );
 }
