--- src/css/Service.css
@@ -1,78 +1,0 @@
-.swing-button {
-  animation: swing 1.5s ease-in-out infinite;
-  width: 150px;
-}
-.circle-icon {
-  width: 100px;
-  height: 100px;
-  background-color:grey; /* Màu xanh dương */
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: white;
-  font-size: 2em;
-  position: relative;
-}
-.circle-text p {
-  font-size: 1.1em;
-  color: #333;
-  max-width: 500px;
-  margin: 0;
-  text-align: center;
-}
-
-.icon-animation {
-  animation: bounce 1.5s infinite alternate;
-}
-
-@keyframes bounce {
-  0% {
-    transform: translateY(0);
-  }
-  100% {
-    transform: translateY(-10px);
-  }
-}
-
-.custom-orange {
-  background-color: orange;
-  color: white;
-}
-@keyframes swing {
-  0% {
-    transform: translateX(0);
-  }
-  25% {
-    transform: translateX(-10px);
-  }
-  50% {
-    transform: translateX(10px);
-  }
-  75% {
-    transform: translateX(-10px);
-  }
-  100% {
-    transform: translateX(0);
-  }
-}
-.info-card {
-  background-color: #f9f9f9;
-  border-radius: 8px;
-}
-
-.info-title {
-  font-weight: bold;
-  color: #ff6a00; /* Màu cam cho tiêu đề */
-}
-
-.info-list {
-  list-style-type: none;
-  padding-left: 0;
-}
-
-.info-list li::before {
-  content: "✔️";
-  color: #ff6a00; /* Màu cam cho dấu check */
-  margin-right: 8px;
-}
