+++ src/page/home/Home.jsx
@@ -1,15 +1,12 @@
-import { useEffect } from 'react';
+import AutoScrollToTop from "../../utils/AutoScollToTop"
 
 export default function Home(){
-    useEffect(() => {
-        window.scrollTo(0, 0);
-      }, []);
+    AutoScrollToTop()
     return (
         <div>
         <h1 className="text-orange">Hello</h1>
         <h1 className="text-orange">Hello</h1>
         <h1 className="text-orange">Hello</h1>
-        
         </div>
     )
 }
