+++ src/utils/AutoScollToTop.jsx
@@ -1,0 +1,8 @@
+import { useEffect } from 'react';
+
+export default function AutoScrollToTop(){
+    return useEffect(() => {
+        window.scrollTo(0, 0);
+    }, []);
+}
+
